[#122] ref: exec: refing dirs
