--- src/exec/search_path.c
@@ -1,83 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   search_path.c                                      :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: yuotsubo <yuotsubo@student.42.fr>          +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/10/13 03:08:02 by yuotsubo          #+#    #+#             */
-/*   Updated: 2024/12/13 14:00:37 by yuotsubo         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "minishell.h"
-
-bool	is_directory(const char *path)
-{
-	struct stat	buf;
-
-	if (stat(path, &buf) < 0)
-		return (false);
-	return (S_ISDIR(buf.st_mode));
-}
-
-static bool	is_dot_path(const char *filename)
-{
-	if (ft_strcmp(filename, ".") == 0 || ft_strcmp(filename, "..") == 0)
-		return (true);
-	return (false);
-}
-
-char	*prepare_path(char *path, char *end, char *value, const char *filename)
-{
-	ft_bzero(path, PATH_MAX + 1);
-	end = ft_strchr(value, ':');
-	if (!end)
-		ft_strlcpy(path, value, PATH_MAX);
-	ft_strlcpy(path, value, end - value + 1);
-	ft_strlcat(path, "/", PATH_MAX);
-	ft_strlcat(path, filename, PATH_MAX);
-	return (end);
-}
-
-char	*do_serch_path(char *value, char *path, const char *filename)
-{	
-	bool	is_permission_denied;
-	char	*end;
-
-	while (*value)
-	{
-		end = prepare_path(path, end, value, filename);
-		if (access(path, F_OK) != 0)
-		{
-			if (end == NULL)
-				return (NULL);
-			value = end + 1;
-		}
-		if (access(path, X_OK) >= 0)
-			return (xstrdup(path));
-		is_permission_denied = true;
-		if (end == NULL)
-			break ;
-		value = end + 1;
-		continue ;
-	}
-	if (is_permission_denied)
-		err_exit(filename, "Permission denied", 126);
-	return (NULL);
-}
-
-char	*search_path(const char *filename)
-{
-	char	path[PATH_MAX + 1];
-	char	*value;
-
-	if (*filename == '\0')
-		return (xstrdup(""));
-	if (is_dot_path(filename))
-		return (NULL);
-	value = map_get(gs_env(GET, NULL), "PATH");
-	if (!value)
-		return (NULL);
-	return (do_serch_path(value, path, filename));
-}
