+++ src/exec/wait_pipeline.c
@@ -1,0 +1,42 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   wait_pipeline.c                                    :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: tkitahar <tkitahar@student.42tokyo.jp>     +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/12/13 15:53:42 by tkitahar          #+#    #+#             */
+/*   Updated: 2024/12/13 15:53:55 by tkitahar         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+
+int	wait_pipeline(pid_t last_pid)
+{
+	pid_t	wait_result;
+	int		status;
+	int		wstatus;
+
+	while (1)
+	{
+		wait_result = wait(&wstatus);
+		if (wait_result == last_pid)
+		{
+			if (WIFSIGNALED(wstatus))
+				status = 128 + WTERMSIG(wstatus);
+			else
+				status = WEXITSTATUS(wstatus);
+		}
+		else if (wait_result < 0)
+		{
+			if (errno == ECHILD)
+				break ;
+			else if (errno == EINTR)
+				continue ;
+			else
+				fatal_error("wait");
+		}
+	}
+	return (status);
+}
